refactor(telegram): split commands.ts into commands/ directory
Extract the 1400-line monolithic commands.ts into 6 domain-specific files:
- commands/admin.ts — restart, update, dangerous mode
- commands/menu.ts — menu keyboard, help, agents, clear, buildSkillMessages
- commands/integrate.ts — integrate flow + callbacks
- commands/session.ts — cancel, status, topics, cleanup
- commands/new-session.ts — new session creation flow + state
- commands/index.ts — setupCommands, setupAllCallbacks, re-exports

Each file registers its own grammY callback handlers with distinct prefixes.
External API preserved via re-exports from index.ts.
